requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
apscheduler==3.10.4
openai==1.51.2
chromadb==0.5.15
tiktoken==0.7.0
